requirements.txt
black==23.1.0
click==8.1.3
colorama==0.4.6
imageio==2.26.0
mypy-extensions==1.0.0
Nuitka==1.4.8
numpy==1.24.2
ordered-set==4.1.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
PySide6==6.4.2
PySide6-Addons==6.4.2
PySide6-Essentials==6.4.2
shiboken6==6.4.2
tomli==2.0.1
zstandard==0.20.0
